[ParkingLot][jericah]refactor: changed getPositionRate return data type from int to double

+++ src/main/java/com/oocl/cultivation/ParkingLot.java
@@ -64,7 +64,7 @@ public class ParkingLot {
         return parkingTicketCarMap.size();
     }
 
-    public int getPositionRate() {
+    public double getPositionRate() {
         return getPositionsAvailable() / getParkingLotCapacity();
     }
